Fix default author build errors
The pages for posts without a listed author were throwing errors in the Actions checker. The template for post authors only checks for the existence of a link on listed authors. When there is no author, the else statement just grabs the name and first listed link from site.social. That results in an <a> tag with no reference.