Change up the homepage committee image styling.
Layouts derived from "default" run their main body through a formatting liquid script called refactor-content. On pages that are not the "home" layout, that encases all images in an <a> tag for the click-to-show-full-img thing. The class for that has a small margin at the top and bottom, but none on the sides. That looks strange. On the "home" layout, the script instead encases the image in a div with ".preview-img". That looks really cool. I duplicated the "page" layout and renamed it to "home", removing the internal checks for "page". The only thing left to get it perfect was a slight adjustment to the aspect ratio style.